+++ tests/Help.Tests.ps1
@@ -1,6 +1,6 @@
 $ModuleName = 'RobocopyPS'
 
-BeforeAll {
+BeforeDiscovery {
     $ModuleName = 'RobocopyPS'
     $ModuleBase = Split-Path -Parent $PSScriptRoot
     $ModulePath = Join-Path $ModuleBase -ChildPath $ModuleName
@@ -8,8 +8,9 @@ BeforeAll {
     # Removes all versions of the module from the session before importing
     Get-Module $ModuleName | Remove-Module
     Import-Module $ModulePath -PassThru -ErrorAction Stop
-}
 
+    $commands = Get-Command -Module $ModuleName -CommandType Cmdlet, Function  # Not alias
+}
 
 Describe "$ModuleName" {
 
@@ -58,6 +59,21 @@ Describe "Invoke-Robocopy" {
     foreach ($Option in $RobocopyOptions) {
         It "$Option has a parameter or alias" -TestCases @{'o' = $Option ; 'all' = $allParameterNameAndAlias} {
             $o | Should -BeIn $all
+        }
+    }
+}
+
+Describe "$ModuleName" -ForEach $commands {
+    BeforeDiscovery {
+        $commandName = $_.Name
+        $help = Get-Help $_.Name
+        $examples = $help.Examples.example
+        $parameters = $help.parameters.parameter
+    }
+
+    Context "$commandname parameters" -Foreach @{param = $parameters} {
+        It "<_.Name> Should not have auto filled description from PlatyPS" -TestCases $param {
+            $param.description | should -Not -BeLike "*{{ Fill * Description }}*"
         }
     }
 }